refactor: detangle diagram by duplicating references
The code is a mess, so make two instances of things if there are two instances in code.  Make the diagram easier to read.. but takes more time to notice things are duplicated.  The code should duplication well though.

Hindsight.... the dcat structure is messy because things are done in set operations in code, not necessarily code iterations... so ... the logic is harder to ascertain.  This should probably be refactored to follow the xml diagram.. but the current design stays truer to the code, so I'm torn **shrug**